Drop credentials input and clean up CI jobs
This removes the deprecated "credentials" input and unifies many of the CI jobs into less test runners to save resources. It also moves cleanup into a recurring cronjob that deletes services that are more than 4 hours old instead of relying on cleanup to occur as part of the integration test suite.