Update error messages to better handle long commands
When running this action with large number of environment variables the
resulting gcloud command gets very long, and if it fails the error
message visible in github actions output is cut off before it gets to
the actually useful bit with the error message, making it impossible to
tell what went wrong.

This adjusts the message the put the command at the end. It's likely
less interesting since it's constructed from inputs the user provides
anyway.